+++ datagrump/controller.hh
@@ -11,8 +11,8 @@ private:
   bool debug_; /* Enables debugging output */
 
   double the_window_size;
-  int64_t skewed_lowest_owt;
-  int64_t lowest_rtt;
+  double skewed_lowest_owt;
+  double lowest_rtt;
 
 public:
   /* Public interface for the congestion controller */
